feat: Implement Pokemon application
feat: Implement Pokémon application with the following features: - Fetch Pokémon data from the API and display it on the Home page . - Extract detailed information for each Pokémon from the API and render it on the Pokémon Details page. - Integrate pagination to enhance navigation within the application.